# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  normal: "normal",
  warning: "warning",
  critical: "critical",
} as const;

export interface Vitals {
  id: number;
  patientId: number;
  heartRate: number;
  spo2: number;
  bpSystolic: number;
  bpDiastolic: number;
  temperature: number;
  respiratoryRate?: number | null;
  timestamp: string;
}

export interface Patient {
  id: number;
  name: string;
  age: number;
  mrn: string;
  gender?: string | null;
  ward: string;
  bedNumber?: string | null;
  diagnosis?: string | null;
  status: PatientStatus;
  riskScore: number;
  admittedAt: string;
  latestVitals?: Vitals | null;
}

export type PatientDetail = Patient & {
  attendingPhysician?: string | null;
  notes?: string | null;
};

export interface CreatePatient {
  name: string;
  age: number;
  mrn: string;
  gender?: string;
  ward: string;
  bedNumber?: string;
  diagnosis?: string;
  attendingPhysician?: string;
}

export interface IngestVitals {
  heartRate: number;
  spo2: number;
  bpSystolic: number;
  bpDiastolic: number;
  temperature: number;
  respiratoryRate?: number;
}

export interface Bed {
  id: number;
  ward: string;
  bedNumber: string;
  occupied: boolean;
  patientId?: number | null;
  patientName?: string | null;
  patientStatus?: PatientStatus | null;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  patientId: number;
  patientName: string;
  type: string;
  severity: AlertSeverity;
  message: string;
  value?: number | null;
  timestamp: string;
  acknowledged: boolean;
}

export type RiskPredictionRiskLevel =
  (typeof RiskPredictionRiskLevel)[keyof typeof RiskPredictionRiskLevel];

export const RiskPredictionRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  severe: "severe",
} as const;

export interface RiskPrediction {
  patientId: number;
  riskScore: number;
  riskLevel: RiskPredictionRiskLevel;
  factors: string[];
  recommendation?: string;
}

export interface DashboardSummary {
  totalPatients: number;
  criticalCount: number;
  warningCount: number;
  normalCount: number;
  totalBeds: number;
  occupiedBeds: number;
  activeAlerts: number;
  recentAlerts: Alert[];
}

export type GetPatientVitalsParams = {
  limit?: number;
};

export type AllocateBedBody = {
  patientId: number;
};

export type ListAlertsParams = {
  limit?: number;
};