# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  patientId: number;
  patientName: string;
  type: string;
  severity: AlertSeverity;
  message: string;
  value?: number | null;
  timestamp: Date;
  acknowledged: boolean;
}