# lib/api-zod/src/generated/types/bed.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from "./patientStatus";

export interface Bed {
  id: number;
  ward: string;
  bedNumber: string;
  occupied: boolean;
  patientId?: number | null;
  patientName?: string | null;
  patientStatus?: PatientStatus | null;
}