# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";

export interface DashboardSummary {
  totalPatients: number;
  criticalCount: number;
  warningCount: number;
  normalCount: number;
  totalBeds: number;
  occupiedBeds: number;
  activeAlerts: number;
  recentAlerts: Alert[];
}