# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from "./patientStatus";
import type { Vitals } from "./vitals";

export interface Patient {
  id: number;
  name: string;
  age: number;
  mrn: string;
  gender?: string | null;
  ward: string;
  bedNumber?: string | null;
  diagnosis?: string | null;
  status: PatientStatus;
  riskScore: number;
  admittedAt: Date;
  latestVitals?: Vitals | null;
}