# lib/api-zod/src/generated/types/patientStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  normal: "normal",
  warning: "warning",
  critical: "critical",
} as const;