# lib/api-zod/src/generated/types/vitals.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Sentinel - Decentralized ICU Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface Vitals {
  id: number;
  patientId: number;
  heartRate: number;
  spo2: number;
  bpSystolic: number;
  bpDiastolic: number;
  temperature: number;
  respiratoryRate?: number | null;
  timestamp: Date;
}